requirements.txt
discord.py
pdfplumber
transformers
torch
